clean up and coalesce in new file
